requirements.txt
Django==2.2.28
djangorestframework==3.11.0
psycopg2==2.9.9
pytz==2024.1
sqlparse==0.5.0
